test(server): extract the shared randomized op model
The three property suites in this package each carried their own copy of
the same I/U/D arbitrary over ids a-d, the same current.json seed, and
(in two of them verbatim, in compactor-randomized inlined) the same
model-apply loop. Widening the op kinds or the id alphabet meant editing
three files, and missing one would leave that suite exploring a narrower
space without any signal.

Move the arbitrary, the seed, the model-apply loop, and the shared keys
into _internal/randomized-model.ts. compactor-randomized keeps its own
reconstructView: it forward-probes past tail_hint, which its idempotence
assertion depends on. No bundler entry reaches the new module and it is
not re-exported through _internal/testing.ts, so it stays test-only.
