+++ packages/server/src/maintenance-safety-randomized.test.ts
@@ -1,93 +1,25 @@
-/* eslint-disable no-underscore-dangle -- `_id` is the locked primary key. */
 import { fc, test } from "@fast-check/vitest";
 import { describe, expect } from "vitest";
+import { logDeleteFloorOf, logSeqStartOf, MemoryStorage, readCurrentJson } from "@baerly/protocol";
 import {
-  CURRENT_JSON_SCHEMA_VERSION,
-  createCurrentJson,
-  type DocumentData,
-  logDeleteFloorOf,
-  logSeqStartOf,
-  MemoryStorage,
-  readCurrentJson,
-} from "@baerly/protocol";
+  applyOps,
+  COLLECTION,
+  CURRENT_JSON_KEY,
+  type Doc,
+  LOG_PREFIX,
+  opArb,
+  PROP_TIMEOUT_MS,
+  reconstructView,
+  seedCurrentJson,
+} from "./_internal/randomized-model.ts";
 import { compact } from "./compactor.ts";
-import { foldLogEntriesOnto, walkLogRange } from "./log-walk.ts";
 import type { InternalMaintenanceOptions } from "./maintenance-options.ts";
 import { runScheduledMaintenance } from "./maintenance.ts";
 import { createObservabilityContext, runWithContext } from "./observability/context.ts";
 import { loadSnapshotAsMap } from "./snapshot.ts";
 import { Writer } from "./writer.ts";
 
-const PROP_TIMEOUT_MS = 600_000;
-const CURRENT_JSON_KEY = "app/x/tenant/t/manifests/tickets/current.json";
-const LOG_PREFIX = "app/x/tenant/t/manifests/tickets";
-const COLLECTION = "tickets";
 const ORPHAN_CONTENT_KEY = `${LOG_PREFIX}/content/${"f".repeat(32)}.json`;
-
-type Doc = DocumentData & { _id: string; v: number };
-
-const opArb = fc.oneof(
-  fc.record({
-    kind: fc.constant("I" as const),
-    id: fc.constantFrom("a", "b", "c", "d"),
-    v: fc.integer({ min: 0, max: 99 }),
-  }),
-  fc.record({
-    kind: fc.constant("U" as const),
-    id: fc.constantFrom("a", "b", "c", "d"),
-    v: fc.integer({ min: 0, max: 99 }),
-  }),
-  fc.record({ kind: fc.constant("D" as const), id: fc.constantFrom("a", "b", "c", "d") }),
-);
-
-const reconstructView = async (storage: MemoryStorage): Promise<Record<string, DocumentData>> => {
-  const read = await readCurrentJson(storage, CURRENT_JSON_KEY);
-  if (read === null) {
-    throw new Error("current.json missing");
-  }
-  const base =
-    read.json.snapshot === null
-      ? new Map<string, DocumentData>()
-      : await loadSnapshotAsMap(storage, read.json.snapshot, COLLECTION);
-  const tail = await walkLogRange(
-    storage,
-    LOG_PREFIX,
-    logSeqStartOf(read.json),
-    read.json.tail_hint,
-  );
-  foldLogEntriesOnto(base, tail, { collection: COLLECTION });
-  return Object.fromEntries(base);
-};
-
-const applyOps = async (
-  writer: Writer,
-  model: Map<string, Doc>,
-  ops: ReadonlyArray<{ kind: "I" | "U" | "D"; id: string; v?: number }>,
-): Promise<void> => {
-  for (const op of ops) {
-    if (op.kind === "I") {
-      if (model.has(op.id)) {
-        continue;
-      }
-      const doc: Doc = { _id: op.id, v: op.v! };
-      await writer.commit({ op: "I", collection: COLLECTION, docId: op.id, body: doc });
-      model.set(op.id, doc);
-    } else if (op.kind === "U") {
-      if (!model.has(op.id)) {
-        continue;
-      }
-      const doc: Doc = { _id: op.id, v: op.v! };
-      await writer.commit({ op: "U", collection: COLLECTION, docId: op.id, body: doc });
-      model.set(op.id, doc);
-    } else {
-      if (!model.has(op.id)) {
-        continue;
-      }
-      await writer.commit({ op: "D", collection: COLLECTION, docId: op.id });
-      model.delete(op.id);
-    }
-  }
-};
 
 // Internal-only seams (the public `MaintenanceOptions` cannot express
 // either). `gc.graceMillis: 0` makes the GC slice's marks due in-pass
@@ -110,15 +42,7 @@ describe("runScheduledMaintenance — never deletes a live object", () => {
     "after a pass: reader view unchanged, live objects survive, stale log reclaimed",
     async ({ ops1, ops2 }) => {
       const storage = new MemoryStorage();
-      await createCurrentJson(storage, CURRENT_JSON_KEY, {
-        schema_version: CURRENT_JSON_SCHEMA_VERSION,
-        snapshot: null,
-        tail_hint: 0,
-        log_seq_start: 0,
-        writer_fence: { epoch: 0, owner: "test", claimed_at: "" },
-        snapshot_bytes: 0,
-        snapshot_rows: 0,
-      });
+      await seedCurrentJson(storage);
       const writer = new Writer({ storage, currentJsonKey: CURRENT_JSON_KEY, options: {} });
       const model = new Map<string, Doc>();
